Proyecto-DFS/requirements.txt
grpcio
grpcio-tools
protobuf
fastapi
uvicorn
python-jose